Adding the OSCInjector to Reaktion
Currently added: OSCInjector - Use an OSC float value from 0 to 1 to
control a Reaction script.

It requires the asset UniOSC but a talented coder could redo this to
work with their own OSC asset choice. In this the OSCAddress and Values
are created as a list from the OSCInjector scripts on start and then
these are initiated on Awake.

Currently working:
OSCInjector functions with Reaktor script when on Use OSC for Audio
Level is enabled, and the OSC Injector is bound to a Script.
Addresses must be defined before run.

Currently not working:
Using a curve and smoothing properties on the incoming OSC data.
Using OSC to control the Gain and Offset and their subsequent options